Improved collision system, added/modified a few flag setting functions, got basic physics manager working

+++ hvn3/include/hvn3/core/Properties.h
@@ -31,7 +31,7 @@ namespace hvn3 {
 			// Flow Properties
 
 			bool FreezeWhenLostFocus;
-			float FrameRate;
+			double FrameRate;
 			int MaxFrameSkip;
 			bool FixedFrameRate;
